requirements.txt
streamlit==1.25.0 
plotly==5.15.0
pandas==2.0.3
numpy==1.24.4
folium==0.14.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.13.0
pillow==9.5.0
pip==23.1.2
inflection==0.5.1
